[refactor] patch_vllm (#3306)
