+++ swift/llm/infer/infer_engine/utils.py
@@ -5,17 +5,21 @@ import os
 import re
 from collections import OrderedDict
 from concurrent.futures import ThreadPoolExecutor
+from contextlib import contextmanager
 from dataclasses import dataclass
+from functools import partial
 from itertools import repeat
 from queue import Queue
 from typing import List, Optional
 
 import torch
+import torch.distributed as dist
 from packaging import version
 from transformers import GenerationConfig, LogitsProcessor
 from transformers.generation.streamers import BaseStreamer
 
 from swift.llm.model.register import fix_do_sample_warning
+from swift.utils import get_device
 from ..protocol import RequestConfig
 
 
@@ -342,3 +346,45 @@ def patch_lmdeploy(load_weights=False):
     TurboMindInstance.__origin_init__ = TurboMindInstance.__init__
     TurboMindInstance.__init__ = __init_ins__
     TurboMindInstance._create_model_instance = _create_model_instance
+
+
+@contextmanager
+def patch_vllm():
+    from vllm.distributed.parallel_state import GroupCoordinator
+    from unittest.mock import patch
+    world_size_patch = patch('torch.distributed.get_world_size', return_value=1)
+    profiling_patch = patch(
+        'vllm.worker.worker.Worker._assert_memory_footprint_increased_during_profiling', return_value=None)
+    __origin_init__ = GroupCoordinator.__init__
+
+    def __init__(self, group_ranks, *args, **kwargs):
+        rank = dist.get_rank()
+        if [rank] not in group_ranks:
+            group_ranks.append([rank])
+        return __origin_init__(self, group_ranks, *args, **kwargs)
+
+    GroupCoordinator.__init__ = __init__
+
+    try:
+        with world_size_patch, profiling_patch:
+            yield
+    finally:
+        GroupCoordinator.__init__ = __origin_init__
+
+
+def patch_npu_vllm(vllm_device: str):
+    if isinstance(vllm_device, int):
+        vllm_device = get_device(vllm_device)
+    device_type = vllm_device.split(':')[0]
+
+    @contextlib.contextmanager
+    def new_group_context():
+        original_new_group = torch.distributed.new_group
+        try:
+            torch.distributed.new_group = partial(original_new_group, use_local_synchronization=True)
+            torch.npu.mem_get_info = partial(torch.npu.mem_get_info, device=vllm_device)
+            yield
+        finally:
+            torch.distributed.new_group = original_new_group
+
+    return new_group_context() if device_type == 'npu' else contextlib.nullcontext()
